requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic[email]>=2.9.0
motor>=3.6.0
python-dotenv>=1.0.0
